+++ app/home/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("home", __name__)
+
+from . import views
