+++ app/prescriptions/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("prescriptions", __name__)
+
+from . import views
